fix: moves validation
